Fix error in length encoding
The encoded length for a string < 32 bytes is stored in the LAST byte of
the variable slot; this was incorrectly using the entire low-word of the
slot. As a result, the last bit is examined to determine if a it is
stores a <32 byte string, or >=32; Odd number (and varSlot() 0x1)
indicates >= 32, while even number implies <32 byte string.

- Update layout comment/doc
- Inspect odd/even-ness of storage slot to choose encoding type
- exit if asked to store a zero length string